Refactor moves
