+++ examples/npt_equilibration.py
@@ -12,8 +12,10 @@ from md.ensembles import PotentialEnergyModel, NPTEnsemble
 
 from md.thermostat.utils import sample_velocities
 
-from md.barostat.utils import get_group_indices, simulate_npt_traj
+from md.barostat.utils import get_group_indices
 from md.barostat.moves import MonteCarloBarostat
+
+from md.utils import simulate_npt_traj
 
 from fe.free_energy import AbsoluteFreeEnergy, construct_lambda_schedule
