+++ md/thermostat/moves.py
@@ -1,0 +1,4 @@
+from md.moves import MonteCarloMove
+
+class UnadjustedMDMove(MonteCarloMove):
+    raise(NotImplementedError)
